organise code for add license
Some codes in ibcpnet.c is completely self-written, not applying
the copyright declaration in the first several license of that file,
and it is not related to the toolkit function for low-level RDMA call.

These codes are moved to ibbuf.c, and applies to some other license,
which will be declared in the next commit.